# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: HomeStock
  description: '

    A self-hosted home inventory system.


    Every response uses the same envelope: `{"data": ..., "error": null}`.

    On failure `data` is null and `error` holds a code and a message.


    The server runs every AI model on the CPU. There is no GPU. The AI routes

    return a job instead of a result when inference takes longer than

    `HS_AI_INLINE_TIMEOUT` seconds.

    '
  version: 0.1.0
paths:
...
  /api/items:
    get:
      tags:
      - Items
      summary: List items with filters, sorting, and pages.
      operationId: list_items_api_items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Full text search terms.
          title: Q
        description: Full text search terms.
      - name: location_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
      - name: include_sublocations
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Sublocations
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: tag_id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
              format: uuid
          - type: 'null'
          title: Tag Id
      - name: is_lent
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Lent
      - name: condition
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(new|good|fair|poor)$
          - type: 'null'
          title: Condition
      - name: owner
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: The person who owns it.
          title: Owner
        description: The person who owns it.
      - name: warranty_expiring_days
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: Warranty Expiring Days
      - name: sort
        in: query
        required: false
        schema:
          type: string
          description: Prefix with - to reverse.
          default: -created_at
          title: Sort
        description: Prefix with - to reverse.
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Page_ItemRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Items
      summary: Create one item.
      operationId: create_item_api_items_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ItemDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/items/owners:
    get:
      tags:
      - Items
      summary: List the people who own something.
      description: The names that the items already carry. The field is free text.
      operationId: list_owners_api_items_owners_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_list_str__'
      security:
      - HTTPBearer: []
  /api/items/search:
    get:
      tags:
      - Items
      summary: Full text search over name, description, brand, model, and notes.
      description: PostgreSQL ranks the results with the generated tsvector column on the items table.
      operationId: search_items_api_items_search_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Q
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Page_ItemRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/items/bulk:
    post:
      tags:
      - Items
      summary: Create many items at once, after an AI bulk scan.
      operationId: bulk_create_items_api_items_bulk_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemBulkCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_list_ItemDetail__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/items/barcode/{code}:
    get:
      tags:
      - Items
      - Barcode
      summary: Look up a barcode in the public product databases.
      description: This route is not in the original route list. The barcode service needs an entry point,
        and the mobile scanner calls it before it creates an item. It tries Open Food Facts, then UPCitemdb.
      operationId: lookup_barcode_api_items_barcode__code__get
      security:
      - HTTPBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_BarcodeProduct_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Items
      - Barcode
      summary: Look a barcode up and create the item from the result.
      description: 'This route is not in the original route list. It is the second half of the barcode
        flow: the scanner may create the item in one call instead of reading the product and posting it
        back.'
      operationId: create_item_from_barcode_api_items_barcode__code__post
      security:
      - HTTPBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      - name: location_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
      - name: quantity
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Quantity
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ItemDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/items/{item_id}:
    get:
      tags:
      - Items
      summary: Return one item.
      operationId: get_item_api_items__item_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ItemDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Items
      summary: Change an item.
      operationId: update_item_api_items__item_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ItemDetail_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Items
      summary: Delete an item.
      description: This is a soft delete. The row keeps a deleted_at value so that the mobile application
        learns about the deletion on its next sync.
      operationId: delete_item_api_items__item_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Message_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/items/{item_id}/photos:
    post:
      tags:
      - Items
      summary: Upload one or more photographs of an item.
      description: The server resizes the original, generates the thumbnail sizes, and removes the EXIF
        GPS data.
      operationId: upload_item_photos_api_items__item_id__photos_post
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_item_photos_api_items__item_id__photos_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_list_ItemPhotoRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Envelope_ItemDetail_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/ItemDetail'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[ItemDetail]
...
    Envelope_Page_ItemRead__:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Page_ItemRead_'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[Page[ItemRead]]
...
    Envelope_list_str__:
      properties:
        data:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Data
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[list[str]]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ItemBulkCreate:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ItemCreate'
          type: array
          maxItems: 100
          minItems: 1
          title: Items
      type: object
      required:
      - items
      title: ItemBulkCreate
      description: Create many items at once, after an AI scan.
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: A JWT access token.
      scheme: bearer